Fix of mariadb-ownsetup patch. Since we do CMake out-of-source builds, and the SOURCE and BUILD directiories differ, the bug that was present in the patch started to cause a issues.
In the CMake "CONFIGURE_FILE(<input> <output>)" command, the <input> path is treated relative to SOURCE dir, however the <output> patch is relative to BUILD dir.
Since the file has been copied to a wrong location, instead in-place updated as originally intended, the file that was later installed was the original unchanged file.

Resolves: #1873999, #1874446

Fix mariadb-pcdir.patch hunk