Add ignorefiles for xmllint rpminspect check
based on https://lists.launchpad.net/maria-discuss/msg06133.html
discussion are all the invalid xmls (except for Index.xml)
present in the sources on purpose and the xmllint
check can be thus disabled in the CI process.

All of xmlfiles, whose warnings are being supressed with this commit are
being used for testing.

A bug report [https://jira.mariadb.org/browse/MDEV-26905] was created
for the Index.xml file.

Fedora CI picks up the rpmlimspect.yaml for specific package in the
dist-git repo [ref:
https://rpminspect.readthedocs.io/en/latest/configuration.html#rpminspect-yaml
]

Replace xmllint shutdown with ignoring specific xmls in the xml check